+++ src/main/kotlin/com/tezos/lang/michelson/editor/completion/provider/NestedTypeCompletion.kt
@@ -3,15 +3,18 @@ package com.tezos.lang.michelson.editor.completion.provider
 import com.intellij.codeInsight.completion.CompletionParameters
 import com.intellij.codeInsight.completion.CompletionProvider
 import com.intellij.codeInsight.completion.CompletionResultSet
+import com.intellij.codeInsight.completion.CompletionType
 import com.intellij.codeInsight.lookup.LookupElementBuilder
 import com.intellij.util.ProcessingContext
 import com.tezos.lang.michelson.lang.MichelsonLanguage
 
 class NestedTypeCompletion : CompletionProvider<CompletionParameters>() {
     override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
-        for (name in MichelsonLanguage.TYPES_COMPLEX) {
-            val item = LookupElementBuilder.create(name).withTypeText("type", true);
-            result.addElement(item)
+        if (parameters.completionType == CompletionType.BASIC) {
+            for (name in MichelsonLanguage.TYPES_COMPLEX) {
+                val item = LookupElementBuilder.create(name).withTypeText("type", true);
+                result.addElement(item)
+            }
         }
     }
 }