+++ src/main/kotlin/com/tezos/lang/michelson/lang/macro/ConditionalMacroMetadata.kt
@@ -9,7 +9,7 @@ class ConditionalMacroMetadata : MacroMetadata {
                 "IF_SOME")
     }
 
-    override fun staticMacroName(): Collection<String> = NAMES
+    override fun staticNames(): Collection<String> = NAMES
 
     override fun validate(macro: String): Pair<String, Int>? {
         return if (macro in NAMES) {